Flask-model/requirements.txt
Flask
torch
torchvision
pillow
numpy